Change num type

+++ pro-3a/parcial1/prueba.c
@@ -1,11 +1,11 @@
 #include <stdio.h>
 int main(){
-	//int num;
-	float num;
+	int num;
+	//float num;
 
 	printf("Ingrese un numero: ");
-	scanf("%f",&num);
-	printf("%.7f", num);
+	scanf("%d",&num);
+	printf("%.7d", num);
 
 	return 0;
 }
